requirements.txt
allure-pytest==2.9.45
jsonpath==0.82
loguru==0.6.0
paramiko==2.9.2
pytest==7.0.1
PyYAML==6.0
requests==2.27.1
xlrd==2.0.1
xlwt==1.3.0
yagmail==0.15.277
mitmproxy==7.0.4
pymysql==1.0.2